requirements.txt
crewai
python-dotenv
langchain-community
pypdf